Refactoring `add task` functionality
Move out `add task` component from `tab one screen` component and make
it a reusable component (namely `TaskItems.tsx`, which renamed from
`TaskList.tsx`). Planning to use `TaskItems.tsx` in `tab two screen` to
display only completed task.
