--- components/TaskList.tsx
@@ -1,53 +1,0 @@
-import React from 'react';
-import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
-
-const TaskList = ({task}) => {
-  return (
-    <View style={styles.item}>
-      <View style={styles.itemLeft}>
-        <TouchableOpacity style={styles.square}>
-        </TouchableOpacity>
-        <Text style={styles.itemText}>
-          {task}
-        </Text>
-      </View>
-      <View style={styles.circular}>
-      </View>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  item: {
-    backgroundColor: '#ffffff',
-    margin: 15,
-    borderRadius: 10,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  itemLeft: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    flexWrap: 'wrap',
-  },
-  itemText: {
-    maxWidth: '80%',
-  },
-  square: {
-    width: 24,
-    height: 24,
-    backgroundColor: '#55bcf6',
-    borderRadius: 5,
-    marginRight: 15,
-  },
-  circular: {
-    width: 12,
-    height: 12,
-    borderColor: '#55bcf6',
-    borderWidth: 2,
-    borderRadius: 5,
-  },
-});
-
-export default TaskList;
